+++ src/utils/themes.js
@@ -1,0 +1,99 @@
+export default [
+    {
+        name: "Aqua",
+        id: 0,
+        themes: {
+            dark: {
+                background: "#121212",
+                primary: "#F06292",
+                secondary: "#3b88d5",
+            },
+            light: {
+                background: "#f2f2f2",
+                primary: "#F06292",
+                secondary: "#64B5F6",
+            },
+        },
+    },
+    {
+        name: "Matsuri",
+        id: 1,
+        themes: {
+            dark: {
+                background: "#121212",
+                primary: "#FCC281",
+                secondary: "#A6C449",
+            },
+            light: {
+                background: "#f2f2f2",
+                secondary: "#839C35",
+                primary: "#FCC281",
+            },
+        },
+    },
+    {
+        name: "Korone",
+        id: 2,
+        themes: {
+            // https://coolors.co/a8736d-f4dba2-f4d7cd-93d3ce-e46d8d <-- Korone color swatch
+            dark: {
+                background: "#0D0707",
+                primary: "#B07975",
+                secondary: "#D57E3D",
+            },
+            light: {
+                background: "#f2f2f2",
+                primary: "#F6D68D",
+                secondary: "#80CDC7",
+            },
+        },
+    },
+    {
+        name: "Rushia",
+        id: 3,
+        themes: {
+            dark: {
+                background: "#121212",
+                primary: "#22C4AC",
+                secondary: "#5783BF",
+            },
+            light: {
+                background: "#f2f2f2",
+                primary: "#9E65A6",
+                secondary: "#F2C9EB",
+            },
+        },
+    },
+    {
+        name: "Fubuki",
+        id: 4,
+        themes: {
+            dark: {
+                background: "#161723",
+                primary: "#FDFDFD",
+                secondary: "#83B0BD",
+            },
+            light: {
+                background: "#FDFDFD",
+                primary: "#317192",
+                secondary: "#83B0BD",
+            },
+        },
+    },
+    {
+        name: "Figaro",
+        id: 5,
+        themes: {
+            dark: {
+                background: "#121212",
+                primary: "#8FB5B7",
+                secondary: "#688BA1",
+            },
+            light: {
+                background: "#F5F5F5",
+                primary: "#688BA1",
+                secondary: "#8FB5B7",
+            },
+        },
+    },
+];