+++ src/views/OrgMusic.vue
@@ -102,6 +102,7 @@
                             v-model="search"
                             :append-icon="icons.mdiMagnify"
                             :label="$t('component.search.searchLabel')"
+                            ref="searchbox"
                             single-line
                             hide-details
                         ></v-text-field>
@@ -177,6 +178,10 @@ export default {
                     ...this.$route.query,
                     page: undefined,
                 },
+            });
+            // eslint-disable-next-line func-names
+            this.$nextTick(function () {
+                this.$refs.searchbox.focus();
             });
         }, 500),
     },